refactor: factor public routes into own file
