+++ lib/moonbeam.js
@@ -7,21 +7,15 @@ const helmet = require('helmet')
 const Joi = require('joi')
 const async = require('async')
 const uuid = require('uuid/v4')
-const dfns = require('date-fns')
-const isValidAccount = require('eos-name-verify')
 
 const {
   sendGenericError,
-  formatCandles,
-  formatTrades
+  log
 } = require('./utils')
 
-const Sunbeam = require('sunbeam')
+const RoutesPub = require('./routes-pub')
 
-const { Api, JsonRpc } = require('eosjs')
-const { JsSignatureProvider } = require('eosjs/dist/eosjs-jssig')
-const fetch = require('node-fetch')
-const { TextDecoder, TextEncoder } = require('util')
+const Sunbeam = require('sunbeam')
 
 const app = express()
 app.use(helmet())
@@ -35,17 +29,14 @@ const schema = {
   }).max(20).unknown(true)
 }
 
-const log = (...args) => {
-  console.error.apply(null, args)
-}
-
 class Moonbeam {
   constructor (conf, plugins) {
     this.conf = conf
 
     this._plugins = plugins
 
-    this.setupRoutes()
+    const pubRoutes = new RoutesPub(conf, plugins)
+    this.setupRoutes(pubRoutes)
 
     this.server = null
     this.plugins = {}
@@ -66,22 +57,6 @@ class Moonbeam {
     this.listenerRegistered = false
 
     this.ws.setMaxListeners(500)
-
-    const fc = this.conf.faucet
-    if (!fc) {
-      return
-    }
-
-    const signatureProvider = new JsSignatureProvider(fc.keyProvider)
-    const rpc = new JsonRpc(fc.httpEndpoint, { fetch })
-    const api = new Api({
-      rpc,
-      signatureProvider,
-      textDecoder: new TextDecoder(),
-      textEncoder: new TextEncoder()
-    })
-
-    this.faucetEos = { api }
   }
 
   connect (cb = () => {}) {
@@ -168,7 +143,7 @@ class Moonbeam {
     async.waterfall(tasks, cb)
   }
 
-  setupRoutes () {
+  setupRoutes (pubRoutes) {
     const { cors } = this.conf
     const cmw = _cors(cors)
 
@@ -177,180 +152,18 @@ class Moonbeam {
     app.post('/history', cmw, this.onPrivateHistoryHttpRequest.bind(this))
 
     // https://api.bitfinex.com/v2/candles/trade:1m:tBTCUSD/last
-    app.get('/v2/candles/:type/last', cmw, this.onCandlesLastHttpRequest.bind(this))
+    app.get('/v2/candles/:type/last', cmw, pubRoutes.onCandlesLastHttpRequest)
     // https://api.bitfinex.com/v2/candles/trade:1m:tBTCUSD/hist
-    app.get('/v2/candles/:type/hist', cmw, this.onCandlesHistHttpRequest.bind(this))
+    app.get('/v2/candles/:type/hist', cmw, pubRoutes.onCandlesHistHttpRequest)
     // https://api-pub.bitfinex.com/v2/trades/tBTCUSD/hist
-    app.get('/v2/trades/:symbol/hist', cmw, this.onPublicTradesHttpRequest.bind(this))
+    app.get('/v2/trades/:symbol/hist', cmw, pubRoutes.onPublicTradesHttpRequest)
 
-    app.get('/time', cmw, this.onTimeHttpRequest.bind(this))
+    app.get('/time', cmw, pubRoutes.onTimeHttpRequest)
 
     const fc = this.conf.faucet
     if (fc) {
-      app.post('/faucet', cmw, this.onFaucetRequest.bind(this))
+      app.post('/faucet', cmw, pubRoutes.onFaucetRequest)
     }
-  }
-
-  isTimeFrameSupported (f) {
-    return [
-      '1M', '6h', '12h', '1D', '7D', '14D', '1M',
-      '1m', '5m', '15m', '30m', '1h', '3h'
-    ].includes(f)
-  }
-
-  checkCandlesReq (type, cb) {
-    if (!type || typeof type !== 'string') {
-      return cb(new Error('ERR_INVALID_PATH'))
-    }
-
-    const framePair = type.replace(/^trade:/, '')
-    const [ frame, pair ] = framePair.split(':')
-    if (!this.isTimeFrameSupported(frame)) {
-      return cb(new Error('ERR_INVALID_TIMEFRAME'))
-    }
-
-    cb(null, [ frame, pair ])
-  }
-
-  onTimeHttpRequest (req, res) {
-    res.status(200).json([ Date.now() ])
-  }
-
-  onFaucetRequest (req, res) {
-    const fc = this.conf.faucet
-    const user = req.body.user
-
-    if (!isValidAccount(user)) {
-      return res.status(500).json({ error: 'ERR_INVALID_USERNAME' })
-    }
-
-    this.faucetEos.api.transact({
-      actions: [{
-        account: fc.contract,
-        name: 'send',
-        authorization: [{
-          actor: fc.account,
-          permission: 'active'
-        }],
-        data: {
-          account: user
-        }
-      }]
-    }, {
-      blocksBehind: 3,
-      expireSeconds: 30
-    }).then((tx) => {
-      res.status(200).json({ ok: true, tx: tx.transaction_id })
-    })
-      .catch((e) => {
-        log(e)
-
-        if (e.json && e.json.error && e.json.error.code === 3050003) {
-          return res.status(500).json({ error: 'ERR_USER_DOES_NOT_EXIST' })
-        }
-
-        res.status(500).json({ error: 'ERR_Q_GENERIC' })
-      })
-  }
-
-  onPublicTradesHttpRequest (req, res) {
-    const symbol = req.params.symbol
-
-    const params = req.query
-    const dStart = new Date(params.start || 1)
-    const dEnd = new Date(params.end || Date.now())
-    const limit = Math.min(+params.limit || 120, 1000)
-    const sort = +params.sort || -1
-
-    const stmt = {
-      symbol: symbol,
-      created_at: { $lte: dEnd, $gte: dStart }
-    }
-
-    const trades = this.plugins.publicTradesDb.db.collection('trades')
-    trades
-      .find(stmt)
-      .limit(limit)
-      .sort({ created_at: sort })
-      .project({ _id: 0, section: 0, created_at: 0, pair: 0, symbol: 0 })
-      .toArray((err, data) => {
-        if (err) {
-          log(err)
-          return res.status(500).json({ error: 'ERR_Q_GENERIC' })
-        }
-
-        const fentries = formatTrades(data)
-        res.status(200).json(fentries)
-      })
-  }
-
-  onCandlesHistHttpRequest (req, res) {
-    const type = req.params.type
-    const params = req.query
-
-    this.checkCandlesReq(type, (err, data) => {
-      if (err) {
-        return res.status(403).json({ error: err.message })
-      }
-
-      const nowDt = Date.now()
-      const maxEnd = dfns.endOfDay(nowDt).getTime()
-
-      const end = +params.end || maxEnd
-      const start = +params.start || 1
-
-      if (!start || !end || start >= end || end > maxEnd) {
-        return res.status(403).json({ error: 'time_interval: invalid' })
-      }
-
-      const limit = Math.min(+params.limit || 120, 3000)
-      const stmt = {
-        key: type,
-        t: { $lte: end, $gte: start }
-      }
-
-      const candles = this.plugins.publicTradesDb.db.collection('candles')
-      candles
-        .find(stmt)
-        .limit(limit)
-        .sort({ t: -1 })
-        .toArray((err, data) => {
-          if (err) {
-            log(err)
-            return res.status(500).json({ error: 'ERR_Q_GENERIC' })
-          }
-
-          const fentries = formatCandles(data)
-          res.status(200).json(fentries)
-        })
-    })
-  }
-
-  onCandlesLastHttpRequest (req, res) {
-    const type = req.params.type
-    this.checkCandlesReq(type, (err, data) => {
-      if (err) {
-        return res.status(403).json({ error: err.message })
-      }
-
-      const candles = this.plugins.publicTradesDb.db.collection('candles')
-      const stmt = {
-        key: type
-      }
-
-      candles
-        .find(stmt, { limit: 1 })
-        .sort({ t: -1 })
-        .toArray((err, entries) => {
-          if (err) {
-            log(err)
-            return res.status(500).json({ error: 'ERR_Q_GENERIC' })
-          }
-
-          const fentries = formatCandles(entries)
-          res.status(200).json(fentries[0])
-        })
-    })
   }
 
   verify (reqId, payload, cb) {
